FitSuiteWidget -> FitSessionWidget (files renamed)
